fix(rider): guard ride-in-progress map against invalid coords (white screen)
PR #1936 fixed the searching screen (driver-arriving) white-screen crash by
validating pickup/dropoff coords before mounting react-native-maps — a
non-finite lat/lng native-crashes the map (white screen, uncatchable by the JS
ErrorBoundary). The in-trip screen (ride-in-progress) still had the same
unguarded pattern, so the identical crash could strike once the ride started.

Extract the validator into rider-app/utils/rideMapCoords.ts (shared) and:
- driver-arriving.tsx now imports it instead of a local copy (dedup)
- ride-in-progress.tsx validates coords, only mounts MapView when all four are
  finite, and feeds the validated values to initialRegion, markers, and the
  route origin/destination; otherwise shows the existing loading placeholder.

Coords are also coerced from numeric strings, matching the merged helper.

Co-developed with Claude Code (claude-sonnet-4-6)
